+++ main.py
@@ -4,7 +4,7 @@ gm_microscope = MyMicroscope("Maria Goeppert Mayer",
     gm_camera = MyCamera("Camera", "28939"),
     gm_spectrometer_vis = MySpectrometer("Spectrometer-VIS", "QEP06226", dark_correction=True),
     gm_spectrometer_nir = MySpectrometer("Spectrometer-NIR", "NQ51B1981"),
-    gm_white_light = MyWhiteLight("White light", "xxxx"),
+    gm_white_light = MyWhiteLight("White light", "68800377"),
     gm_laser_750 = MyLaser("Laser 750nm", "xxxx"),
     gm_laser_550 = MyLaser("Laser 550nm", "xxxx"),
     gm_filter_12 = MyFilter("Filter 12", "xxxx"),
@@ -18,5 +18,5 @@ lm_microscope = MyMicroscope("Lise Meitner",
     lm_filter_12 = MyFilter("Filter 12", "xxxx"),
     lm_platform = MyPlatform("Platform", "xxxx"))
 
-app = MyApp("V1.0.3", (1200, 700), gm_microscope, lm_microscope)
+app = MyApp("V1.0.4", (1200, 700), gm_microscope, lm_microscope)
 app.mainloop()